smart_cradle_server_v2/requirements.txt
Flask
Flask_SQLAlchemy
pymysql
paho-mqtt
numpy
opencv-python-headless
gunicorn
cryptography
